main/linux-rpi: fix config prompt during cross-compilation
The ARCH parameter was missing from the syscconfig call, which resulted in
incomplete configuration during cross-compilation and broke the build with
an interactive "Restart config..." prompt when building _dev packages.

Also adds a check to ensure that _carch is set, so that an unexpected CARCH
value will be caught early and not trigger similar unwanted behavior.