community/gnome-session: do not depend on gnome-shell
GNOME Shell is simply the default shell that gnome-session runs
if not otherwise specified. However, it is not a hard dependency.
As gnome-shell is already marked as a dependency
for gdm and gnome packages, removing the dependency does not have
any consequence in the GNOME stack, while allowing for more flexibility
on programs that use gnome-session to run a non-default shell (e.g: phosh)