main/mkinitfs: handle compressed kernel modules
When determining which firmware needs to be includes in the initramfs
mkinitfs searchs for kernel modules using the filename *.ko and so
does not cater for compressed modules (and therefore does not include
their required firmware in the initramfs).

Change the search pattern to deal with both uncompressed and
compressed modules.

Fixes #13188.