main/openssh: use a more sensible default $PATH value
Workaround for $PATH order issue with SSH, this is due to the fact that
path is now appended instead of overwritten in /etc/profile and thus the
order of the default SSH path is not changed. By supplying a better
default SSH path value we can workaround this issue for now.

Merge requested by @ncopa.

See #13017