community/dnsdist: re-enable -luajit subpackage again
This looks like it was accidentally removed in
9cf0aa8f7bd2f3640a50bc7da4b655fe7c83a90d in this commit the -luajit
subpackage is removed entirely instead of just enabling it for aarch64
again. Since the commit only removes the subpackage but not the split
function this looks like a mistake to me.

This fixes a dependency error (see #13173).

See also f774aa10980364a1659d29f3da489f3889c38ca7