community/postgresql-bdr: move from main
The last release was in 2017, i.e. this project is dead and we can
hardly provide support that is expected from packages in the main
repository. I wanted to remove it, but Leonardo Arena objected:

> These packages are currently deployed in hundreds of installations in
> a large corporate network which is contributing to the Alpine project.
> So it will be detrimental for Alpine to remove them.

For this reason, we keep it in aports and just move to the community
repository.

The only package that depends on postgresql-bdr is
postgresql-bdr-extension.