community/fennel: fix build with options="!check"
Without this patch, fennel.lua is only build because it is a prerequisite
of the test target. If tests are disabled with options="!check" (e.g. on
the riscv64 builder) then fennel.lua is not build at all and package()
fails.

This commit fixes this issue by explicitly building fennel.lua.