main/hylafaxplus: shipped hfaxd.conf has now .default extension
hfaxd.conf is not in /etc dir, so at each upgrade this file
is overwritten rather than renamed with .apk-new.

With this patch, the shipped hfaxd.conf has .default extension.
In this way the hfaxd.conf file is never overwritten with the
upgrades.

the openrc init.d file checks if there's no hfaxd.conf and alert the user
following the same approach that is done with setup.cache.

Fixes #13085