main/gcc: re-add mmap patch for libgo
I removed this patch previously because I assumed that this was supposed
to be a workaround for compilation error. However, some further testing
of gcc-go on 32-bit architectures actually revealed that this patch
fixes a run-time error. Without this patch, the libgo runtime is
incapable of allocating memory on 32-bit architectures, though it works
fine without on 64-bit architectures. I updated the patch description to
clarify its use and re-added it.