testing/agg: build with sdl12-compat-dev
I have no idea what this is used for. Grepping for agg-dev gives me
zero results and the original commit message said that it's
"predominantly" used by gnash, the GNU Flash Player thing, which no
longer seems to be in aports anyway, so I don't know how to test
whether this works. Maybe it would be wise to drop it.

Part of https://gitlab.alpinelinux.org/alpine/aports/-/issues/12739