Revert "community/gcc-cross-embedded: enable support for additional RISC-V ext."
While this builds fine locally, it seems to cause some issues on the
builders. Two libstdc++/libsupc++ variants which are present in $pkgdir
locally are not present on the builders for some reason. Unfortunately,
the build logs is half a gigabyte in size and thus it is very difficult
to debug this in a timely manner. As such, revert for now and revisit
this issue later.

See https://gitlab.alpinelinux.org/alpine/aports/-/merge_requests/28155#note_196716

This reverts commit d07d6c6ec49e4f69004d9e5a1e87e25b4310ab74.