main/util-linux: remove -dev dependency on util-linux
This was introduced in dcc5965f6f4fabd096b4df6cdd01e4f206341f64.
Unfortunately, no comment was added explaining why this dependency is
necessary. The -dev subpackage only provides library files, header
files, and pkg-config files. None of these should require any programs
from util-linux.

If you are reading this in the future and this commit turns out to be a
problem, i.e. util-linux is actually needed as a dependency of
util-linux-dev, then please add a comment to the APKBUILD.

Approved by @Leo (the author of the aforementioned commit).