community/icingaweb2: remove optional dependencies
This commit removed the dependency on gd and imagick, which reduces
the installed size of this aport (including dependencies) by about
100mb.

These php modules are only used when you run an pdf-export in icingaweb,
a feature which is only rarely used in the wild (afaik).

If required, these modules can be installed manually afterwards without
problems to provide the pdf-export feature.