community/libcec: add udev rule to make USB adapter part of input group
The Pulse-Eight HDMI-CEC adapter by default appears as /dev/ttyACM0,
owned by the "dialout" user with 660 permissions. However this device is
used as an input device so it makes more sense to be owned by the
"input" group. This also allows any user part of that group to use the
device without having to be part of the dialout group, which is useful
for at least Plasma Bigscreen and probably also for Kodi