main/coreutils: remove testcase from fix-chmod-exit-code.patch
We don't run tests anyhow at the moment for coretuils (these are
disabled with options=!check) but the test case changes modify a build
system configuration file and thus require partially regenerating the
autotools Makefiles. However, this only seems to be supported with
Automake 1.16.4, not with Automake 1.16.5 (which we package atm):

	configure.ac:38: error: version mismatch.  This is Automake 1.16.5,
	configure.ac:38: but the definition used by this AM_INIT_AUTOMAKE
	configure.ac:38: comes from Automake 1.16.4.

By removing the build system modification for the test case from
fix-chmod-exit-code.patch, we can circumvent this error and fix the
coreutils build failure on the builders.