community/docker: upgrade to 20.10.12
This is mostly just a version bump over 20.10.12, as we do not bundle the 'scan' cli plugin -- nor do we package it -- which was updated to find Log4j issues.

In effect this simply satisfies the automated message from pkgs.alpinelinux.org that the package is out of date.

For more information, see https://docs.docker.com/engine/release-notes/#201012