community/obexd-enhanced: move OBEXD out of main/bluez
OBEXD has now PBAP capabilities to share an Evolution's addressbook
with Bluetooth clients such as a car multimedia system.
However, it requires evolution-data-server-dev dependencies which
cannot be moved into main.

This package provides the same OBEXD as in the main BlueZ package
but with enhanced capabilities such as PBAP support.
Users can opt-in for this package which replaces OBEXD provided by the
bluez-obexd package.