community/rtkit: add polkit rules
Add a rule to allow members of the polkit group enable realtime
scheduling for their processes. This can be useful for users of
pipewire that are not running it in an (e)logind session, e.g. users
of seatd or headless setups.

fixes https://gitlab.alpinelinux.org/alpine/aports/-/issues/13432