main/linux-lts: refactor kernel configs
Refactor our kernel configs to be a minimal diff from the defconfig,
using `savedefconfig`. This makes our kernel config only inlude the
difference from the default kernel config.

Move the oldconfig to build phase instead of prepare. This makes it
easier to update the kernel configs for all architectures when needed
and it is more consisten with other packages which does the configure
part in the buidl phase instead of prepare.

Also introduce prepareupdate and updateconfigs helper functions to make
it easier to update the kernel and kernel configs.

Fixes https://gitlab.alpinelinux.org/alpine/aports/-/issues/12933