testing/python2: move from community
There are no more packages in community/ which require Python 2. As
such, we can now move the python2 package to testing as well. This
allows us to ship the upcoming Alpine 3.16 release without Python 2
support (see https://gitlab.alpinelinux.org/alpine/tsc/-/issues/40).

As soon as we have fixed the remaining Python 2 packages in testing/ we
can remove Python 2 from Alpine entirely. See #12740 for the current
progress on these efforts. See also #11577 for the history on this.