community/phoc: build embedded wlroots
Phoc requires a few out of tree patches to wlroots, and trying to
maintain separate package forks of wlroots in aports to apply these
patches is a maintenance burden and doesn't always work as expected
(e.g. sometimes apk ignores it and installs an unpatched wlroots,
breaking phoc for users)

This basically gives up the fight for now and embeds purism's wlroots
fork w/ their patches to support phoc/phosh.