ci: remove before_script
Due to recent changes in git, the owner of the repository files matter.
In CI, the repository is cloned by root, while the CI jobs run as a
non-root user. This causes git to fail with an error message.

This was already fixed in the ci build script, but the `before_script`
is run before that, and now starts to fail because the updated git
version is part of the image itself.

Because we fetch the commits in the build script anyway, lets just
remove the before_script.