community/kyotocabinet: disable project's build-time compiler optimizations
This project uses -march=native by default, which causes problems when
the package is built on a system with a newer ISA than the target CPUs.
--disable-opt causes it to use the default -march in gcc, which should
be something safe (right?)

In this particular case, it fixes an issue where the swpa instruction
was being included by gcc, causing a SIGILL on CPUs (like the cortex
A53) where this instruction is *not* supported. See:
https://github.com/rinigus/osmscout-server/issues/345