community/chez-scheme: move from testing
Has been in testing/ for almost two years and works quite well in
my experience. Furthermore, the package is now available for all
architectures supported by upstream and tests are enabled and pass.
As such, I don't see any reason to keep it in testing/ any longer
and it is also needed as a dependency of other packages (e.g. idris2)
and hence would be neat to have in community/ to also be able to move
these other packages to community/ in the future.