main/busybox: tweak udhcpc timeout defaults
Change the udhcp discover timeout from 3 seconds to 1 and increase
retries to 15.

It may take a second from ip link set up to interface gets carrier. This
means that during boot the first try will always fail and we would wait
from first timeout. By chainging it from 3 seconds to 1, we reduce the
time til we get the ip address to 1 second.

Previously we had timout=3 and retries=5 due to
https://gitlab.alpinelinux.org/alpine/aports/-/issues/3105 so we also
adjust the retries so that we in total retry for up to 15 seconds.