testing/pypy-bootstrap: rename from pypy-stage0
works around a bug in abuild where apk info --who-owns on the .so files
returns -stage0 (the package name of this bootstrap), which then adds it
to the depends of pypy, which makes it uninstallable as they cannot be
pulled in at the same time. like this, it will instead add
pypy-bootstrap, which will be resolved by itself.