main/grub: prevent duplicated menu entries
If a /etc/grub.d/ template file is locally customised/modified then
if the Grub package is later updated the stock version of that file
will be placed in /etc/grub.d/ with a .apk-new suffix. Whenever
grub-mkconfig is next triggered (e.g. after kernel package update)
it will make use of both the modified file and the stock file so
resulting in the creation of duplicate menu entries.

This patch prevents that by ignoring any *.apk-new files in the
template directory.