main/eggdrop: upgrade to 1.9.2
removed eggdrop.post-install, /usr/bin/eggdrop-installer
does not exist, and there is no install= line in the
APKBUILD, so the post-install was not being used anyway.

removed gseen.mod.patch and logs2html.mod.patch as they are very
old and not part of the official release tarball, they are not
really patches anyway, but full source files, so if anyone needs
them, it's probably better to add them to source= instead.