community/wolfssl: disable bigcache for default build
When I ran collectd with collectd-write_riemann built against
riemann-c-client with WolfSSL, it surprised me that the virtual size
of the collectd process (VIRT in htop) jumped to ~700 MiB (from a
few megabytes). I tried recompiling wolfssl without --enable-bigcache,
then running it again and the VIRT size of the collectd process
returned back to usual numbers.