main/linux-lts: add support for rk3399
If a config got removed, it's because savedefconfig detected we're using
the default value. They end up having the value I describe in this
commit message though.

To get an interactive console when drivers are missing, all of these need
to be `y`
- CONFIG_SERIAL_8250_DW
- CONFIG_REGULATOR_FIXED_VOLTAGE
- CONFIG_CLK_RK3399

Ethernet:
- CONFIG_STMMAC_PLATFORM: it defaults to `m`
- CONFIG_ROCKCHIP_PHY

Useful for peripherals:
- CONFIG_I2C_RK3X
- CONFIG_SPI_ROCKCHIP
- CONFIG_PWM_ROCKCHIP
- CONFIG_PHY_ROCKCHIP_USB
- CONFIG_ROCKCHIP_THERMAL
- CONFIG_ROCKCHIP_SARADC
- CONFIG_PCIE_ROCKCHIP_HOST: I choose `y` because other platforms do the
  same

Power management:
- CONFIG_ROCKCHIP_PM_DOMAINS: `m` is not supported

Security and Virtualization:
- CONFIG_ROCKCHIP_IOMMU: `m` is not supported
- CONFIG_ROCKCHIP_IODOMAIN

Others:
- CONFIG_RANDOM_TRUST_CPU: still the same value, but `y` is now the default