community/librsvg: move rsvg-convert to subpackage
This tool has ~5 MiB and it's not needed for librsvg.so.

Fedora provides it in package librsvg-tools, Debian in librsvg-bin,
Void Linux in librsvg-utils, SUSE in rsvg-convert. Since there's only
a single utility/tool, rsvg-convert, and it's like this for a very long
time, I think that rsvg-convert is the best option for the subpackage
name.