main/abuild: backport change for gomodcache
since we added gocache=$srcdir everywhere, this is now needed to let
anything build at all, as otherwise the readonly directory cannot be
deleted. the actual merge request is preferable
(https://gitlab.alpinelinux.org/alpine/abuild/-/merge_requests/134), but
it would need deployment- this sets it as default and deploys it
instantly, and is not a breaking change for anyone (unless they were
relying on a cache to be readonly?)

if an abuild release was tagged and deployed sometime before, this would
not be needed.