community/fwupd: handle no plugins installed
When no plugins are installed, the plugin directory is not available
because all plugins are split of in subpackages. However, fwupd daemon
will access this directory to load any installed plugins. It aborts
when the directory does not exist. Therefore, create an empty directory
during build to handle the case when the user hasn't installed any
plugins. Also add a merged patch to avoid an assert when no plugins are
installed. This patch can be dropped when the next fwupd release is
available.