main/alpine-base: add mdev-conf to alpine-base dependancy list
With the various restructuring changes in Alpine Edge around
mdev/mdevd/udev I have found during testing on a system being converted
to eudev/udev use that whenever "setup-devd udev" does a
"apk add eudev udev-init-scripts udev-init-scripts-openrc" this not
only automatically triggers the removal of the busybox-mdev-openrc
package but also the removal of mdev-conf.

mdev-conf contains various mdev configuration related files that are
required on all Alpine systems regardless of which of mdev/mdevd/udev
is in use - if mdev-conf is removed then any time mkinitfs rebuilds the
initramfs file (such as when triggered by a kernel package update) the
resultant initramfs will have NO mdev support and so the machine will
fail to boot.

This change ensures that mdev-conf is always a dependancy.