main/gcc: build libgccjit separately
it requires --enable-host-shared- enabling this for all of gcc incurs a
small percentile performance penalty, so Just Don't and if' it out into
its own block.

this does complicate the build a little more- but the other alternative
is to make another aport, and then it would need to follow more or less
the same tarball/patchset, have more moving parts/places to fix flags,
etc.. this is the easier option, reusing the existing configuration.

also, remove -j1 from install to speed it up a little