community/mate-{screensaver,session-manager}: build with elogind support
In MATE 1.24 a elogind compile option was introduced to
mate-session-manager and mate-screensaver.

When installing packages elogind and polkit-elogind, among others WiFi
can be turned on/off with media keys and a suspend button becomes
available in the shutdown dialog.

If elogind isn't installed, this commit has no impact on the behaviour.