community/xorg-server: fix software-based rotation via modesetting
It is already fixed upstream almost a year ago, but there seems to
be no release in sight, so cherry pick it.
It has the most influence on ebook readers, several have a landscape
display and cannot work in portrait mode without that patch, since
refreshes do not work properly.
postmarketOS supports several ebook readers.