community/phosh: execute squeekboard directly instead of going through osk-wayland
I am not sure what the point of this ever was. In Alpine,
osk-wayland does not exist. Downstream in postmarketOS it is a
script that checks for the presence of squeekboard and then
executes squeekboard, and if it does not exist, it defaults to
virtboard which as far as I can tell never existed in Alpine and
has long been dropped from postmarketOS. As such, do away with this
needless indirection and execute squeekboard directly.