main/gcc: add post-install message for gcc-go
We currently do not provide security updates for libgo, i.e. the
packages of the Go standard library. For this (and other) reasons,
it is currently discouraged to use gcc-go for purposes other than
bootstrapping community/go. The added post-install messages seems
to be the right place to document this pitfall.