community/gtk4.0: add patch to support copy-paste on devices without mouse
For devices with a touchscreen and no mouse, copy-paste from GTK 4.0 programs
under Wayland does not work, because GTK 4.8 only considers pointer events
when fetching the last event serial and not touch events.

This is being discussed upstream. Meanwhile, this change adds a patch to make
it use a different function that does consider touch events.

Ref: alpine/aports#14179