testing/tic-80: don't build with pro enabled
The licence technically allows us to do this (TIC-80 is free software,
after all), but I don't think it's right. The BUILD_PRO build option is
meant to build the paid version of TIC-80 that's sold to support
development and is not intended to be enabled in redistributed builds
like this. As such, I think we should appreciate that TIC-80 upstream
is keeping the paid version 100% free software rather than going the
free core route, and disable this option here out of respect.