community/ctags: move from main
ctags is a complicated piece of software with a lot of parsers written
in C. As such, there is a high possibility of bugs occurring during
processing of untrusted input. ctags upstream doesn't do longterm
support releases. As such, I don't think we can realistic support
this software in a safe way for up to two years.