testing/guake: rename tarball
had the same name but we fetch a different source entirely in the last
commit intentionally, as the release tarball is broken and is missing
files to build. funnily, upstream claims you can't use this as the
"build tool" requires git history, but it does not require this at all,
as it's always a single env var away to set manually without it.
python...