community/firefox*: Remove hardcoded path from desktop entry
A common pattern to write wrappers for applications (e.g.: to run
Firefox inside a sandbox), is to place an executable with the same name
in /usr/local/bin/ or ~/.local/bin/ or any other path with higher
precedence in $PATH.

Firefox's desktop entry is one of the very few that includes an absolute
path to the binary. This has a slight downside: it doesn't use wrappers
which follow the above pattern. OTOH, using an absolute path doesn't
seem to bring any upsides (I suspect this approach exists mostly due to
historical reasons).

The safe-mode desktop entry remains unchanged, since that one does not
have a hardcoded path.