community/emacs: explicitly install man and info pages in doc()
Instead of installing everything and then removing everything except man
and info pages. This results in an identical package which the exception
that usr/include/emacs-module.h and usr/share/metainfo/emacs.metainfo.xml
are removed from the -doc packages. IMHO both of these files do not
belong in a -doc subpackage in the first place and are also provided
by the -nox, -x11, and -gtk subpackage.