community/ripgrep: remove jemalloc-dev, cleanup
the size difference between system jemalloc and tree-built is about 10%,
but there is nothing else using the actual system jemalloc so it ends up
actually being bigger. further, ld_preload of another alloc breaks the
executable when it's dynamically linked to jemalloc. this way, it's safe
to have a preload of another alloc (even if it's not in use given the
jemalloc symbol renaming)

also remove the redundant patch, and ensure the jemalloc is built with
16k pages