testing/openocd-riscv: Resolve conflict with OpenOCD
Allow parallel installation with openocd or openocd-git by installing
data in to `usr/share/openocd-riscv` instead of `usr/share/openocd` and
and `-riscv` suffix to binaries. This is useful for users who prefer
working with upstream OpenOCD for e.g. ARM targets, but need RISC-V
patches not upstreamed yet when working with RISC-V targets.