main/ifupdown-ng: add onlink option to default routes
Add the "onlink" next-hop option to the default route. This is required
for situations like some VPS/Cloud Providers where they provide a /32
IPv4 address or a /128 IPv6 address and obviously the router to use as
the gateway for the default route is not in the same subnet.

The ifupdown package has been adding the "onlink" option for the past 7
years when Debian added it:

https://salsa.debian.org/debian/ifupdown/-/commit/8b7bca9597d2f75670b182f0fc149cdbaec3544c

This patch has been submitted upstream as
https://github.com/ifupdown-ng/ifupdown-ng/pull/207.