main/binutils: add -z text by default
makes the linker fail on textrels directly.

we already check for textrels with scanelf- but this crashes in the
link that generates them, which makes it a bit easier to debug why they
occur.

it does, however, need an additional workaround to permit when it cannot
be fixed (adding -Wl,-z,notext to ldflags) on top of options=textrels