community/entr: upgrade to 5.3
The C unit test suite (which has been previously invoked via
`make test`) was removed and replaced by the regression tests
(previously `make regress`). These didn't pass on the builders
before due to lack of a TTY and several bashisms in the test script.
To fix that, run the tests in script(1) and execute the test script
with /bin/bash instead of /bin/sh.