community/networkmanager: add networkmanager-dnsmasq subpackage
NetworkManager uses DBus to configure dnsmasq but the regular dnsmasq in
Alpine does not contain DBus support. The dnsmasq-dnssec-dbus package
provides this feature, so depend on it through a subpackage.
Without this, NetworkManager will never be able to start dnsmasq instances.
Inform users when they install NetworkManager that they should install
the subpackage if they want to make use of NetworkManager dnsmasq
features.