community/tint2: upgrade to 17.1.3
This new version is a fork of the original:
https://gitlab.com/o9000/tint2

The author stated:
"The final release of tint2 is 17.0.2.
The code is frozen and no more feature requests are accepted."

The new author states:
"I had to adopt tint2 from o9000 account rather automatically
after it was suddenly sentenced to end-of-life by previous
maintainer without warning. "

Some Linux distribution has adopted this version:
- Crux
- Gentoo
- LiGurOS
- Termux