main/linux-headers: upgrade to 6.1.11, implement rsyncless install
this is a bit of a hack- however it drops the dependency on `rsync` to
install the headers.

since they are just a tree of `.h` (and we drop the other files by
default), the output is the same.

since rsync depends on a lot of things, and linux-headers is a base
dependency for a lot of things, this eases the bootstrap dependency
chain.

in this case, it allows zstd to use a different build system without
causing a dependency loop[0]; but keeping file-only installations like this
more dependency-free is always nice.

[0]: rsync -> zstd -> meson -> python -> linux-headers -> rsync