main/rng-tools: start rngd after procfs
rngd reads /proc/sys/kernel/random/poolsize at startup and falls back
to 4K if the file was not present.

If /proc is not mounted yet by the time rngd starts, and the real pool
size is smaller (which is likely on low end devices with low memory),
then rngd will keep trying to feed entropy to linux in a tight loop and
consume CPU for no benefit.