community/qtchooser: remove
this is a broken hack that does not work for build scripts-
it doesn't work for finding qt6 versions of things at all, for instance

things should:
- call qmake/qmake-qt5 for qt5
- call qmake6 for qt6 (same for lrelease6, ..)

for qt6, everything usually finds it via cmake.
there are also the /usr/lib/qt$version/bin fallbacks.
there is no point to have a wrapper for `qmake` to dynamically pick the
version, things have to explicitly say what they want and not expect
`qmake` to be 6.

fortunately, nothing calling `qmake` `lrelease` et al expect them to be
6, so this is not used.