community/gnome-control-center: add reapply patch
GNOME Control Center did not reapply connection settings through
NetworkManager when the user changed it. Only when the user toggled the
network device OFF/ON, the settings were applied. It was missing a
reapply call to NetworkManager which was added in MR 1642 [1].
This patch is merged for GNOME 44 but backporting the patch to fix this
bug already.

Only the functional patch is backported, the clean up patches in the MR
are not backported to keep the patch small until GNOME 44 is released.

[1] https://gitlab.gnome.org/GNOME/gnome-control-center/-/merge_requests/1642