community/qt5-qttools: subsplit libraries
closes #14639

the previous change to install the main symlinks causes the main package
with all the libraries to depend on the -dev package too. that in turn
pulls in a chain of -devs.

since things link to the libraries, split them to pull them in directly.

should look at this again for the symlinks..