community/kodi: upgrade to 20.0
this moves around the executables a bit- they don't have to be in usr/
lib/kodi, and dropping the provides=$pkgver lets them be installed in
parallel. we keep -x11 as the default, and also move the `kodi` binary
to it, as all it can do is launch the kodi-x11 backend.