community/ffmpeg: re-add -O3
the testing in e9fa4208e00a9b80e477e6445a28705364f66292 was too simple;
the encoder being used was not in ffmpeg, it was an external library.
so, no optimisations really mattered in that case, as all ffmpeg did was
call codec functions and only some bare-minimum de/muxing.

where these optimisations matter, is where ffmpeg itself does processing.
this consists of:
- networking stuff
- hwaccel submission stuff
- codecs implemented in ffmpeg itself
- filter related things

for something like a complex filtergraph in
https://www.gariany.com/2020/08/ffmpeg-step-by-step-retro-video-filter/
it changes the execution time from 23 seconds to 20 seconds on a sample
video file with the same commandline and procedure. this is because
ffmpeg itself is doing `complex` stuff in such a scenario.

such a 10+% performance in ffmpeg itself increase (and that is without
removing the external libx264 re-encoding part from the metrics.. so the
actual gain is much bigger in ffmpeg itself) is more than worth it for
the size gain.