community/lxcfs: modprobe requires -a for multiple modules
'modprobe fuse autofs4' will pass autofs4 as an argument to fuse
module loading, which is not what was intended here.
Using modprobe with -a properly loads both modules, or skips if
they were built-in, and fails if one of those was neither
built-in nor a module.

Not actually using lxc so not tested, found by manually looking at the
PR looking at autofs or fuse: both actually should be loaded
automatically by the kernel when a pre-created device is opened,
but alpine does not seem to have a way of pre-creating such devices,
so keeping the modprobe here.

Link: https://github.com/alpinelinux/aports/pull/4991