community/phoc: Fix black screen on quick power button presses
With phoc 0.25.0 on PP/PPP the screen goes black (but on) and stays
stuck that way when the power button is pressed two or more times in
quick succession (see [1] for the upstream bug report). The patch
comes from the upstream MR to fix the bug [2].

Upstream maintainer Guido Günther reports it working on Librem 5 in
the bug discussion, I tested it successfully on my PinePhone.

[1] https://gitlab.gnome.org/World/Phosh/phosh/-/issues/900
[2] https://gitlab.gnome.org/World/Phosh/phoc/-/merge_requests/428