main/clang15: install to llvm prefix
this looks like a mess at first glance.

the reason it looks so messy is that cmake requires things to be present
in their original locations-
it is not valid to install files via cmake, then move them after. things
can only be symlinked or cmake discovery will fail later as they are not
present and the location is hardcoded.

that aside, this fixes an issue- projects usually expect the clang
include directory to match the llvm include directory. that is, they
except them to be installed to the same prefix.

usually this isn't that hard to patch, but this also affects things like
mesa rusticl in the future.

as a bonus, you can now install `clang 14` and `clang15` at the same time,
as the former no longer has a `clang` binary, just clang-14.

utilising all the older tools would require adding the
/usr/lib/llvm*/bin/* to path.

for latest-clang, they are all symlinked to bin (this needs careful
checking every upgrade).

i tested that everything still works with this (analyzer, scan build,
the compilers..),
but there might be a regression /somewhere/.