testing/openocd-git: Add udev and cmd_openocd subpkg
The udev rules subpackage function was added, but the subpkg was never
added to `subpackages`, so it was never actually build. This fixes the
issue.

In addition the `openocd-git-cmd-openocd` subpackage is added that just
adds a symlink from `/usr/bin/openocd` to `openocd-git. This will make
life easier for users who only want to use the git version of
OpenOCD without the release version installed as well.