main/knot: loosen /run directory permission
Knot DNS 3.2.5 changed socket permissions to allow any user to
communicate with the DNS server. For this to be usable, any user must be
able to access a socket in the runtime directory.

Search permission should be sufficient for this purpose, as anything
trying to connect to a socket should already know the path it is looking
for.