community/fzf: add vim as dependency for fzf-vim
fzf-vim has a post-install trigger that executes vim. Normally it's
exected that this subpackage is automatically pulled in when having vim
and fzf installed.

But if you install this package explicitly without having vim installed,
the post-install script will fail.

Add vim as dependency as installing this package without vim does not
make a lot of sense.

Fixes: #14741