main/postfix: upgrade to 3.8.0, remove redis support
the redis support was added in
https://gitlab.alpinelinux.org/alpine/aports/-/merge_requests/33818
but this was never accepted upstream and no longer applies onto the
source (conflicts).

not sure why we ever added this when upstream does not merge it,
aports is not for feature development.

upstream thread: https://www.mail-archive.com/postfix-devel@postfix.org/msg00956.html