community/pandoc: embed data files using --constraint
The `embed_data_files` flags is only valid for `pandoc`, not for
`pandoc-cli`. For this reason, we need to pass this flag when
building `pandoc` as a dependency using cabal. This can be achieved
using the `--constraint` flag. Contrary to the solution proposed in
69e1c77a3ee93b34d13e754ab8baab7f157ba74e this does not require us to
build from a GitHub source tarball and still allows us to use `cabal
install` as usual in `package()`.

This partially reverts commit 69e1c77a3ee93b34d13e754ab8baab7f157ba74e.

Fixes #14814