scripts/mkimg.base.sh: remove haveged package
With changes to the "random" code in recent kernels there seems little
need for haveged in general. This script adds the haveged package to
Alpine images but its init.d script never appears to be enabled
anywhere (neither in the image nor by setup-alpine) and so the
package's presence in the image appears unnecessary.