testing/tridactyl-native: vendor nimble package list
Different mirrors seem to ship different package lists at the moment
thus causing the build to fail if a certain mirror is selected. To
workaround that, simply ship a package list with all dependencies
as part of the package.

To-Do: Also ensure that the integrity of the dependencies is verified
by nimble, which doesn't seem to be the case at the moment. That is,
nimble clones the repository and then just checks out a tag without
verifying the integrity of the repository.