community/newlib: add missing nano.specs and syscalls for msp430-elf
For MSP430 no nano.specs is installed for the MSP430 toolchain. Likely
most MSP430 users only install the nano variant and don't build a full
fledged version of newlib. To ease packaging, for consistency, and also
in case some MSP430 users do want a full fledged newlib version,
stick with providing both versions but also provide a simple nano.specs
file to make using the nano variant possible.

Additionally, syscalls where disabled for msp430. The resulted in the
functions provided by the standard C lib such as puts() or printf()
to be defunct, as the hooks integrating with an RTOS such as RIOT
where never called. This patches newlib to also use syscalls with
MSP430.