community/git-annex: rebuild with ghc-9.4.5
This requires some dependency hackery as the previous base-compat
version is not compatible with the new base version of GHC 9.4.5 but
feed is not compatible with the new base-compat version. To workaround
that, allow a newer base-compat version for feed using --allow-newer.
This can hopefully be removed in the future.