main/gcc: improve flag passing
(and remove pkgrel from version spec since that makes it require to
rebuild the kernel every time pkgrel is bumped)

the _FOR_TARGET flags apply to the generated libraries,
and the BOOT_ flags apply to the compiler.

without them, e.g. LDFLAGS are not passed, so dt_relr was missing from
gcc libraries.

be consistent and pass them all. also set -g0 to save a tiny bit of
build space and memory, since we're not using any debug data. it's still
generated with -dbg added.

sadly, this still doesn't pass the ldflags to gcc proper, just the
libraries, but it's a start

and, move libstdc++exp.a to libstdc++-dev

verified bootstrap still works