community/heirloom-doctools: avoid conflicts with common packages
The heirloom-doctools provides several small binaries which are
intimately tied together. Unfortunately, these binaries are also
provided by other common and packages (e.g. coreutils, util-linux-misc
and groff). For this reason, it was previously not possible to have, for
example, util-linux-misc and heirloom-doctools installed simultaneously.

Since the aforementioned packages are quite popular this is a common
annoyance. To workaround this issue, install the heirloom-doctools
binaries and man pages to non-standard directories and provide a
/etc/profile.d configuration file to add these directories to the
PATH and MANPATH.

This approach is also taken by other downstream packagers [1].

[1]: https://github.com/openbsd/ports/blob/master/textproc/heirloom-doctools/pkg/README