community/inkscape: ship tutorials in extra subpackage
Since abuild >= 3.11.5, abuild emits a warning if the -doc subpackage
exceeds 2 MiB. This is the case for inkscape due to the tutorials
which are ~70 MiB in size. For this reason, it makes sense to
ship these tutorials in a separate subpackage.

This fixes the abuild 3.11.5 warning.

See https://gitlab.alpinelinux.org/alpine/tsc/-/issues/16