main/hunspell-en: split into per-dialect subpackages
This package currently provides all dictionaries for all English
dialects/variants. When installing it, applications will detect all
eleven dictionaries and show all of them as an available dictionaries.

This hurts usability substantially; in Firefox I'd expect to see four
dictionaries to quickly chose between the languages that I use, but I
end up seen over a dozen, ten of which are unused English variants. At a
glance, these are also quite similar, so picking for this large menu is
extra irritating.

Split all variants into separate subpackages, to allow the installation
of just the desired dictionaries while skipping the others.

The `hunspell-en` package is now a metapackage that provides all these
new subpackages. This avoid breaking any existing installations as this
package continues providing the same functionality as before.