community/waked: fix qualcomm's rtc errata
The RTC on Qualcomm platforms are apparently read-only and don't reflect
the current time, this commit adds a patch where we use relative instead
of absolute time when writing an alarm to the RTC, as this will work on
sdm845 and similiar. Thank you @caleb, who basically told me the problem
and suggested to just use relative times in waked.