main/gdb: replace patch with upstream patch
The current patch fixing the segfault when inspecting a variable with
an absolute address by adding checks before dereferencing the
`the_bfd_section` member. The upstream patch instead makes sure to just
always provide the member even for variables with absolute addresses.

This replaces the patch with the solution handed in for inclusion in
upstream, see [1] and [2].

[1]: https://sourceware.org/pipermail/gdb-patches/2023-June/200406.html
[2]: https://sourceware.org/bugzilla/show_bug.cgi?id=30431