community/picolibc: fix deps, fix build, rebuild against gcc
Sadly, picolibc depends on the exact version of the embedded GCC cross
compiler it was build with, as it installs its specs into
`/usr/lib/gcc/$target_triple/$gcc_ver/picolibc.specs`.

In order to shoot myself in the foot again next time I bump GCC, the
dependency on GCC now is nailed to the version picolibc was build with.

Finally, CFLAGS and LDFLAGS are cleared prior building, as the host
system flags (for a full Linux OS system) are not all compatible with
the target (tiny bare metal MCU). This fixes compilation issues with
a current Alpine host system.